requirements.txt
Flask
flask-cors
Pillow
requests
python-dotenv
opencv-python
numpy
paddlepaddle